fix: green CI, align versions to 0.5.1 and enforce quality in the pipeline
The GitHub Actions pipeline was failing on two workflows and the repository
carried three different answers to "what version is this?". This closes both,
and adds the CI guards whose absence let the drift accumulate unnoticed.

CI failures

- Metrics Validation / Build workspace: the step compiles with `--locked`, but
  the committed Cargo.lock declared `cli 0.5.0` while the manifests declared
  `0.4.0`. Cargo refused to update the lock and failed before compiling.
- xtask gen-golden --check: `crates/ir/golden/if.ir` still recorded the old,
  invalid emission (`br_cond %f_0`, a temp that is never defined) after the
  lowering fix changed the output. This is the first step of build-and-test and
  would have failed as soon as the IR commits reached CI.
- CI / perf-regression: added in 0ce2e6f and never passed once — 7f4cf33 was its
  first run. The script is rewritten (see below); confirmation depends on the
  next CI run.

Versions

All 11 manifests, Cargo.lock, README and website unified on 0.5.1, matching what
CHANGELOG already documented. Previously: 0.4.0 in the manifests and README,
0.5.0 across the website.

New guards

- `lint` job: cargo clippy --workspace --all-targets --locked -- -D warnings,
  plus cargo fmt --all -- --check. CI ran clippy nowhere, so the "zero warnings"
  claim from b7ba5c3 had regressed silently.
- `--locked` in build-and-test, pinning Cargo.lock to the manifests.
- 53 clippy warnings and 1 hard error cleared; workspace reformatted.

Bugs found while cleaning up

- Recursion overflowed the process stack past depth ~5 on Windows and ~40 on
  Linux, because a debug-build interpreter frame costs ~200 KB of Rust stack.
  The CLI now runs on a dedicated 256 MB-stack thread, so the interpreter's own
  depth-128 guard is reached first. This also fixed cli --test stream_pipeline.
- LSP was broken on Windows: the URI decoder only handled %20/%23/%25, so the
  `file:///c%3A/...` VS Code sends never resolved, and to_file_uri emitted
  `file://\?\C:\...` from canonicalize. Added generic percent-decoding, drive
  letter handling, verbatim-prefix stripping and a URI<->path round-trip test.
- `art add` and the import resolver disagreed on the cache directory — the
  former read HOME with a "." fallback, the latter used dirs::home_dir(). On
  Windows packages landed where the resolver never looked. Both now share
  resolver::cache_dir(), overridable via ARTCODE_HOME.

perf_regression.sh

Rewritten: no python3 dependency, best-of-5 sampling so one unlucky sample on a
shared runner cannot fail the build, stderr preserved instead of discarded, and
the measured JSON printed on failure.

Repository hygiene

- .gitattributes normalizes line endings; without it every Windows checkout
  rewrote the tree to CRLF and reported the whole repo as modified.
- cli/.art-lock untracked — integration tests rewrite it with absolute /tmp
  paths. .art-lock and docs/generated/ added to .gitignore.
- Tests needing POSIX tools (echo, tr, sh) marked #![cfg(unix)].

Docs

Roadmap block T corrected: TTD phase 2 was marked 0/5 despite having been
delivered in a5f4d9f. Verified against the source and restated as 2/5 done plus
3 partial. .kit/checklist.md refreshed against what the code actually contains.

+++ crates/interpreter/tests/stress_actors.rs
@@ -55,7 +55,7 @@ fn test_scheduler_actor_massive_message_stress() {
             // Isso encherá as filas locais sem consumir ainda (exceto as preemptadas).
             for (i, _aid) in actor_ids.iter().enumerate() {
                 for m in 0..messages_per_actor {
-                    let target = actor_ids[(i + (m as usize)) % actors_qty as usize];
+                    let target = actor_ids[(i + m) % actors_qty];
 
                     // actor_send(target, message_value)
                     interp.interpret(vec![Stmt::Expression(Expr::Call {
@@ -73,7 +73,7 @@ fn test_scheduler_actor_massive_message_stress() {
 
             // 3. Forçar o agendador nativo a resolver todo o pool assíncrono enfileirado nas threads C.
             // `run_actors_round_robin(N)` irá drenar as pilhas iterativamente pelo Agendador Nativo VM.
-            interp.run_actors_round_robin((total_messages_expected + 10000) as usize);
+            interp.run_actors_round_robin(total_messages_expected + 10000);
 
             // 4. Verificação final de estado: 
             // Todos os mailboxes devem estar vazios sem panics do Rust Lock (sem deadlock de mutex).
